requirements.txt
openai>=1.12.0
azure-identity>=1.15.0
python-dotenv>=1.0.0
